split up config screen class
