+++ modstitch.ct
@@ -13,3 +13,5 @@ accessible method net/minecraft/client/gui/components/AbstractWidget renderScrol
 accessible method net/minecraft/client/gui/components/AbstractSliderButton setValue (D)V
 accessible field net/minecraft/client/gui/components/EditBox value Ljava/lang/String;
 accessible field net/minecraft/client/gui/screens/Screen renderables Ljava/util/List;
+accessible field net/minecraft/client/gui/screens/Screen minecraft Lnet/minecraft/client/Minecraft;
+accessible field net/minecraft/client/gui/screens/Screen font Lnet/minecraft/client/gui/Font;
