+++ src/main/resources/assets/bigsignwriter/lang/en_us.json
@@ -14,6 +14,7 @@
   "bigsignwriter.font.error.invalidHeight": "Invalid height (%s)!",
   "bigsignwriter.font.error.wrongLineCount": "Either the character '%s' has an incorrect number of lines (%s), or the font's height has been misset (%s)!",
   "bigsignwriter.font.error.unfixedWidth": "Character '%s' does not have a fixed width! %s",
+  "bigsignwriter.symbols": "Symbols",
   "bigsignwriter.config": "Big Sign Writer Config",
   "bigsignwriter.config.openFontsFolder": "Open Fonts Folder",
   "bigsignwriter.config.reloadFonts": "Reload Fonts",
@@ -36,6 +37,8 @@
   "bigsignwriter.config.fontSelectorOpensScrolledUp.desc": "Scrolls the font selector to the top when it's opened to allow double clicking to return to vanilla typing.",
   "bigsignwriter.config.showConfigButton": "Show Config Button",
   "bigsignwriter.config.showConfigButton.desc": "Whether to show the config button next to the font selector",
+  "bigsignwriter.config.showSymbolsButton": "Show Symbols Button",
+  "bigsignwriter.config.showSymbolsButton.desc": "Whether to show the symbols button next to the font selector",
   "bigsignwriter.config.displayFontHeights": "Display Font Heights",
   "bigsignwriter.config.displayFontHeights.desc": "Displays the height of fonts in the font selector",
   "bigsignwriter.config.characterSeparatorOverride": "Character Separator Override",
